Fix FormBuilderField rebuilds by changing parent widget to stateful
키보드가 올라오거나 setState로 인해 상위 위젯이 rebuild되면 새로운 GlobalKey가 만들어지고, 이로 인해 FormBuilderField가 계속해서 새로 빌드되는 문제가 있었음.
그 과정에서 아래와 같은 warning 발생
```
flutter: Warning! Replacing duplicate Field for wrongProblems -- this is OK to ignore as long as the field was intentionally replaced
flutter: Warning! Replacing duplicate Field for reviewProblems -- this is OK to ignore as long as the field was intentionally replaced
flutter: Warning! Ignoring Field unregistration for wrongProblems -- this is OK to ignore as long as the field was intentionally replaced
flutter: Warning! Ignoring Field unregistration for reviewProblems -- this is OK to ignore as long as the field was intentionally replaced
```